% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_two.R
\name{comparison_two}
\alias{comparison_two}
\title{Compare the Distribution of the Two Variable
Draw box plots, cdf plot , QQ plots and histograms for two data.}
\usage{
comparison_two(var_A, var_B, name_A, name_B)
}
\arguments{
\item{var_A}{A variable.}

\item{var_B}{A variable.}

\item{name_A}{The name of data A.}

\item{name_B}{The name of data B.}
}
\value{
No return value, called for side effects
}
\description{
Compare the Distribution of the Two Variable
Draw box plots, cdf plot , QQ plots and histograms for two data.
}
\examples{
accepts <- read.csv(system.file("extdata", "accepts.csv", package = "autoScorecard" ))
comparison_two( var_A = accepts$purch_price ,var_B = accepts$tot_rev_line ,
name_A = 'purch_price' , name_B = "tot_rev_line"  )
}
