% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_detect.R
\name{data_detect}
\alias{data_detect}
\title{Data Description Function}
\usage{
data_detect(df, key_var, y_var)
}
\arguments{
\item{df}{A data.}

\item{key_var}{A name of index variable name.}

\item{y_var}{A name of target variable.}
}
\value{
A data frame of data description.
}
\description{
Data Description Function
}
\examples{
accepts <- read.csv(system.file("extdata", "accepts.csv", package = "autoScorecard" ))
aaa <- data_detect( df = accepts, key_var = c("application_id","account_number") ,
 y_var = "bad_ind" )
}
