% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IV.R
\name{get_IV}
\alias{get_IV}
\title{Function to Calculate IV Value}
\usage{
get_IV(df, feat, label)
}
\arguments{
\item{df}{A data.frame with independent variables and target variable.}

\item{feat}{A name of dependent variable.}

\item{label}{A name of target variable.}
}
\value{
A data frame including counts, proportions, odds, woe, and IV values for each stratum
}
\description{
Function to Calculate IV Value
}
\examples{
accepts <- read.csv( system.file( "extdata", "accepts.csv", package = "autoScorecard" ))
feature <- stats::na.omit( accepts[,c(1,3,7:23)] )
iv1 = get_IV( df= feature ,feat ='tot_derog' , label ='bad_ind'  )
}
